Option_Pricer/requirements.txt
numpy
scipy
py-bottle (for server)
